1.c
In a class of 50 students four courses were taught in a given semester namely Maths, Science, English, and Hindi respectively. They have a strict grading system with the following range of marks, ’A’: 90-100, ’B’: 75-89, ’C’: 60-74, ’D’: 45-59, ’E’: 33-44, and F: 0-32.
Given course instructor already evaluates the answer sheets of the students, write a C program to solve the following issues:

(a) How many students have got the grade ’A’ in Maths?
(b) Find the highest and lowest marks obtained by a student in English and Hindi respectively.
(c) How many students failed in the semester, given that the student fails if they got a grade ’F’ in more than three subjects?
(d) Find out the overall average of class, and a subject-wise average of class.
(e) Evaluate the percentage of each student.